semantic-search-fastapi/requirements.txt
fastapi==0.70.1
uvicorn==0.16.0
openai==0.27.2
pinecone-client==2.0.13
matplotlib
plotly
scipy
pandas
scikit-learn
tiktoken
supabase
PyPDF2==3.0.1
tqdm
